Make TimeLine mutable, support multiple tagged events in EventPlacement
Mutable TimeLine:

Initially I had the idea to have static TimeLine, where events can only
be added but not changed. It turned out its better to keep them mutable
(because this is the general mutwo approach and its hard to fight
against this). To keep it performant the `register` method won't check
for potential overlaps and `EventPlacement` aren't sorted if the user
doesn't ask for it. To fix overlaps we will need a conflict solution
hook in the future.

Support multiple tagged events in EventPlacement:

Sometimes EventPlacement will need to have multiple instruments/players
(to avoid Cagean solitude). Therefore an EventPlacement should support
multiple tags.