interfaces/EventPlacement: Fix for small ranges
If we have start or end ranges with a very small difference
it's possible that we catch a rounding error: the difference
between start & end is smaller than the rounding loss which is
caused by mutwo.core_parameters.configurations.ROUND_DURATION_TO_N_DIGITS.
Because the difference is so small we can simply return only
one value, because the range doesn't really matter anyway.
This patch implements this fix and makes 'EventPlacement' therefore more
robust for corner case usages.