interfaces/TimeLine += .unregister method
We can already add new 'EventPlacement' to a 'TimeLine' via the
'register' method. But before this patch it wasn't possible to remove
'EventPlacement' which were added with the 'register' method. This
commit adds the new 'unregister' method to 'TimeLine', so that we can
remove 'EventPlacement' again if we don't want them anymore.