converters/TimeLineToSimultaneousEvent: Ensure stable tag order
Before this patch, the order of 'TaggedSimultaneousEvent' in the
'SimultaneousEvent' returned by the converter
'TimeLineToSimultaneousEvent' was random (because 'set' type was used
for tag names). Now the order is always the same.

This fixes flaky test 'TimeLineToSimultaneousEventTest.test_convert'.
Before this commit it was good luck if the tag order was the expected
one.